+++ src/mod.ts
@@ -1,8 +1,9 @@
 /**
  * This is the default export of uneventful, which contains the API for jobs and
  * streams, as well as any signals-related APIs that don't depend on the signals
- * framework (e.g. {@link recalcWhen}, which does nothing if the signals framework
- * isn't in use, and doesn't cause it to be imported).
+ * framework (like {@link recalcWhen}() and {@link isObserved}(), which do
+ * nothing if the signals framework isn't in use, and won't cause it to be
+ * imported).
  *
  * For the rest of the signals API, see the
  * [uneventful/signals](uneventful_signals.html) export.